Update shipped_number_of_packs in service layer

+++ server/repository/src/migrations/v2_09_00/add_shipped_number_of_packs_to_invoice_line.rs
@@ -11,7 +11,7 @@ impl MigrationFragment for Migrate {
         sql!(
             connection,
             r#"
-                ALTER TABLE invoice_line ADD COLUMN shipped_number_of_packs REAL;
+                ALTER TABLE invoice_line ADD COLUMN shipped_number_of_packs DOUBLE PRECISION;
             "#
         )?;
